Create smedia
This endpoint will store the uploaded multimedia file on the server and then redirect to the newly created URL.. In this code, we first check if the file part is present in the POST request. If the file part is present, we retrieve the file from the request. If the file is empty or if the filename does not have an allowed extension, we return a 400 Bad Request error. Otherwise, we save the file in the uploads folder with a secure filename.

Finally, we return a 302 Found redirect to the URL of the uploaded file. The browser will then make a GET request to this URL to retrieve the file.

Please note that you should replace /path/to/your/uploads with the actual path to the folder where you want to store the uploaded files.

This code also includes the necessary steps to validate the file and protect against attacks such as MIME type mismatch attacks.